Add: Integreated media uploader, Fix: Code quality changes

+++ backend/src/service/auth.service.ts
@@ -1,17 +1,16 @@
 import bcrypt from "bcrypt";
 import { ZodError } from "zod";
-import HttpStatus from "http-status-codes";
+import { StatusCodes } from "http-status-codes";
 import { NextFunction, Request, Response } from "express";
 import { constructSucessResponse } from "../utils/responseGenerator";
 import {
   verifyRefreshToken,
   generateAccessAndRefreshTokens,
-} from "../utils/jwt";
+} from "../utils/jwtUtils";
+import ApiError from "../utils/apiError";
 import { User } from "../models/user.model";
 import { UserSesion } from "../models/userSession.model";
 import { CreateUserDto, LoginDto } from "../dto/auth.dto";
-import { CustomRequest } from "../entity/auth.entity";
-import ApiError from "../utils/apiError";
 
 export async function createNewUser(
   req: Request,
@@ -28,16 +27,16 @@ export async function createNewUser(
     });
 
     await newUser.save();
-    constructSucessResponse(res, { id: newUser.id }, HttpStatus.CREATED);
+    constructSucessResponse(res, { id: newUser.id }, StatusCodes.CREATED);
   } catch (err: any) {
     if (err instanceof ZodError) {
-      next(new ApiError(HttpStatus.BAD_REQUEST, err.errors[0].message));
+      next(new ApiError(StatusCodes.BAD_REQUEST, err.errors[0].message));
     } else if (
       err.name &&
       err.name === "MongoServerError" &&
       err.code === 11000
     ) {
-      next(new ApiError(HttpStatus.CONFLICT, "User already exists"));
+      next(new ApiError(StatusCodes.CONFLICT, "User already exists"));
     } else next(err);
   }
 }
@@ -52,12 +51,12 @@ export async function loginUser(
     const user = await User.findOne({ email: credentials.email });
 
     if (!user) {
-      throw new ApiError(HttpStatus.NOT_FOUND, "User not found");
+      throw new ApiError(StatusCodes.NOT_FOUND, "User not found");
     }
     const match = await bcrypt.compare(credentials.password, user!.password);
 
     if (!match) {
-      throw new ApiError(HttpStatus.BAD_REQUEST, "Incorrect password");
+      throw new ApiError(StatusCodes.BAD_REQUEST, "Incorrect password");
     }
     const { accessToken, refreshToken } = generateAccessAndRefreshTokens(user);
     await UserSesion.findOneAndUpdate(
@@ -66,11 +65,11 @@ export async function loginUser(
       { upsert: true }
     );
 
-    res.cookie("token", refreshToken, { httpOnly: true, secure: true });
+    res.cookie("rtoken", refreshToken, { httpOnly: true, secure: true });
     constructSucessResponse(res, { accessToken });
   } catch (err) {
     if (err instanceof ZodError) {
-      next(new ApiError(HttpStatus.BAD_REQUEST, err.errors[0].message));
+      next(new ApiError(StatusCodes.BAD_REQUEST, err.errors[0].message));
     } else next(err);
   }
 }
@@ -81,9 +80,8 @@ export async function logoutUser(
   next: NextFunction
 ) {
   try {
-    const { userId } = req as CustomRequest;
-    await UserSesion.findOneAndDelete({ userId });
-    res.clearCookie("token", { httpOnly: true, secure: true });
+    await UserSesion.findOneAndDelete({ userId: req.userId });
+    res.clearCookie("rtoken", { httpOnly: true, secure: true });
     constructSucessResponse(res);
   } catch (err) {
     next(err);
@@ -95,7 +93,7 @@ export async function refreshAccessToken(
   res: Response,
   next: NextFunction
 ) {
-  const refreshToken = req.cookies?.token;
+  const refreshToken = req.cookies?.rtoken;
   try {
     const { userId, email } = await verifyRefreshToken(refreshToken);
     const { accessToken } = generateAccessAndRefreshTokens({
